Save data in local storage
