+++ index.html
@@ -130,5 +130,6 @@
   </footer>
   <script src="./js/script.js" type="module"></script>
   <script src="./js/form-validation.js" type="module"></script>
+  <script src="./js/utility.js" type="module"></script>
 </body>
 </html>
